cpr/src/requirements.txt
fastapi
uvicorn
joblib~=1.0
numpy~=1.20
pandas
gcsfs
xgboost
google-cloud-storage>=1.26.0,<2.0.0dev
google-cloud-aiplatform[prediction] @ git+https://github.com/googleapis/python-aiplatform.git@custom-prediction-routine
